Temp remove icons for prebuilt integrations
